requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==0.28.1
requests==2.31.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 